Move test to proper library
