+++ tests/general/testdata/receivers.yaml
